cpu utilization per thread + mbufs per socket + add in tui

+++ src/trex_defs.h
@@ -37,6 +37,6 @@ limitations under the License.
 
 typedef std::set<uint32_t> flow_stat_active_t;
 typedef std::set<uint32_t>::iterator flow_stat_active_it_t;
-typedef std::vector<std::vector<double>> cpu_util_full_t;
+typedef std::vector<std::vector<uint8_t>> cpu_util_full_t;
 
 #endif
